+++ relay-pii/src/processor.rs
@@ -1,8 +1,8 @@
 use std::borrow::Cow;
 use std::collections::{BTreeMap, BTreeSet};
 use std::mem;
+use std::sync::OnceLock;
 
-use once_cell::sync::OnceCell;
 use regex::Regex;
 use relay_event_schema::processor::{
     self, enum_set, Chunk, Pii, ProcessValue, ProcessingAction, ProcessingResult, ProcessingState,
@@ -137,7 +137,7 @@ impl<'a> Processor for PiiProcessor<'a> {
             match self.process_string(value, meta, state) {
                 Ok(()) => value.push_str(&basename),
                 Err(ProcessingAction::DeleteValueHard) | Err(ProcessingAction::DeleteValueSoft) => {
-                    *value = basename[1..].to_owned();
+                    basename[1..].clone_into(value);
                 }
                 Err(ProcessingAction::InvalidTransaction(x)) => {
                     return Err(ProcessingAction::InvalidTransaction(x))
@@ -374,7 +374,7 @@ fn apply_regex_to_chunks<'a>(
             return;
         }
 
-        static NULL_SPLIT_RE: OnceCell<Regex> = OnceCell::new();
+        static NULL_SPLIT_RE: OnceLock<Regex> = OnceLock::new();
         let regex = NULL_SPLIT_RE.get_or_init(|| {
             #[allow(clippy::trivial_regex)]
             Regex::new("\x00").unwrap()
@@ -469,12 +469,10 @@ mod tests {
     use insta::assert_debug_snapshot;
     use relay_event_schema::processor::process_value;
     use relay_event_schema::protocol::{
-        Addr, Breadcrumb, DebugImage, DebugMeta, Event, ExtraValue, Headers, LogEntry, Message,
+        Addr, Breadcrumb, DebugImage, DebugMeta, ExtraValue, Headers, LogEntry, Message,
         NativeDebugImage, Request, Span, TagEntry, Tags, TraceContext,
     };
-    use relay_protocol::{
-        assert_annotated_snapshot, get_value, Annotated, FromValue, Object, Value,
-    };
+    use relay_protocol::{assert_annotated_snapshot, get_value, FromValue, Object, Val};
     use serde_json::json;
 
     use super::*;
@@ -1336,6 +1334,43 @@ mod tests {
         assert_eq!(
             get_value!(span.data.code_filepath!).as_str(),
             Some("src/sentry/api/authentication.py")
+        );
+    }
+
+    #[test]
+    fn test_ai_token_values() {
+        let mut span = Span::from_value(
+            json!({
+                "data": {
+                    "ai.total_tokens.used": 30,
+                    "ai.prompt_tokens.used": 20,
+                    "ai.completion_tokens.used": 10,
+                }
+            })
+            .into(),
+        );
+
+        let pii_config = serde_json::from_value::<PiiConfig>(json!({
+            "applications": {
+                "$object": ["@password"],
+            }
+        }))
+        .expect("invalid json config");
+
+        let mut pii_processor = PiiProcessor::new(pii_config.compiled());
+        process_value(&mut span, &mut pii_processor, ProcessingState::root()).unwrap();
+
+        assert_eq!(
+            Val::from(get_value!(span.data.ai_total_tokens_used!)).as_u64(),
+            Some(30),
+        );
+        assert_eq!(
+            Val::from(get_value!(span.data.ai_prompt_tokens_used!)).as_u64(),
+            Some(20),
+        );
+        assert_eq!(
+            Val::from(get_value!(span.data.ai_completion_tokens_used!)).as_u64(),
+            Some(10),
         );
     }